simplify registry getters
