+++ src/builtin-addons/core/template-completion-provider.ts
@@ -127,15 +127,8 @@ function isArgumentName(name: string) {
 }
 
 export default class TemplateCompletionProvider {
-  _registry!: IRegistry;
-  _registryVersion = -1;
   get registry(): IRegistry {
-    if (this._registryVersion !== this.project.registryVersion) {
-      this._registry = this.server.getRegistry(this.project.roots);
-      this._registryVersion = this.project.registryVersion;
-    }
-
-    return this._registry;
+    return this.project.registry;
   }
   project!: Project;
   server!: Server;
